style: runAll returns bool, doxygen block comments, auto removal
- Change runAll() return type from int to bool
- Use static_cast<int> at call sites in all 3 test mains
- Replace auto with explicit types in test_fixture.cpp
- Use /*! */ block style for Signal<T> class doxygen
- Convert // comments to //! doxygen in window.hpp
- Minor spacing and blank line cleanup

+++ libs/window/tests/window_tests.cpp
@@ -39,5 +39,5 @@ TEST_CASE(key_event_data)
 
 int main()
 {
-    return TestFixtureLib::runAll();
+    return static_cast<int>(TestFixtureLib::runAll());
 }
